Added limit argument to ansible_parallel.py
Some ansible tasks may produce stderr or stdout messages that exceed the default stream buffer limit. When this happens a string message is appended to the playbook results array, which causes the final `return sum(results)` to bomb out with a `TypeError: unsupported operand type(s) for +: 'int' and 'ValueError'` error.

This change allows a user to specify a new buffer limit as a cli argument to accommodate for such scenarios.